follow api add in serverless yml

+++ src/follow/delete.ts
@@ -7,12 +7,14 @@ const parameter = {
   properties: {
     followedUserIdx: { type: "string" },
   },
-  required: ["user"],
+  required: ["followedUserIdx"],
 } as const;
 
 export const handler = async (event: APIGatewayProxyEventV2WithLambdaAuthorizer<{ [key: string]: any }>) => {
   console.log("[event]", event);
   const { followedUserIdx } = JSON.parse(event.body) as FromSchema<typeof parameter>;
+  if (!followedUserIdx) return { statusCode: 400, body: JSON.stringify({ message: "followedUserIdx is required" }) };
+  
   const userIdx = event.requestContext.authorizer.lambda.idx;
 
   await mysqlUtil.deleteMany("tb_follow", { follower_idx: userIdx, followed_idx: followedUserIdx });
